Moved the `core::support::configuration` module to `core`
