+++ src/main/build.rs
@@ -298,7 +298,7 @@ fn run_bindgen(build_common: &ShadowBuildCommon) {
         .disable_header_comment()
         .raw_line("/* automatically generated by rust-bindgen */")
         .raw_line("")
-        .raw_line("use crate::core::support::configuration::QDiscMode;")
+        .raw_line("use crate::core::configuration::QDiscMode;")
         .raw_line("use crate::host::descriptor::{File, FileSignals, OpenFile};")
         .raw_line("use crate::host::descriptor::socket::inet::{InetSocket, InetSocketWeak};")
         .raw_line("use crate::host::host::Host;")
