+++ src/main/core/mod.rs
@@ -1,3 +1,4 @@
+pub mod configuration;
 pub mod controller;
 pub mod cpu;
 pub mod logger;
@@ -7,6 +8,5 @@ pub mod resource_usage;
 pub mod scheduler;
 pub mod sim_config;
 pub mod sim_stats;
-pub mod support;
 pub mod work;
 pub mod worker;
