--- src/main/core/support/mod.rs
@@ -1,1 +1,0 @@
-pub mod configuration;
