+++ src/main/network/graph/mod.rs
@@ -10,9 +10,7 @@ use log::*;
 use petgraph::graph::NodeIndex;
 use rayon::iter::{IntoParallelIterator, ParallelIterator};
 
-use crate::core::support::configuration::{
-    self, Compression, FileSource, GraphOptions, GraphSource,
-};
+use crate::core::configuration::{self, Compression, FileSource, GraphOptions, GraphSource};
 use crate::network::graph::petgraph_wrapper::GraphWrapper;
 use crate::utility::tilde_expansion;
 use crate::utility::units::{self, Unit};
